Rename SearchCategories -> CategoryList
